ASDK-486 Workaround 0 timescale on erroneous content
The issue was that the Atom parser assumes that it never gets a 0 value
from the SIDX timescale. We had issues with content that had this wrong
value. This workaround falls back the timescale reported in the MDHD box
instead. According to spec, these two values should be identical. This
prevents the issue where a division by zero exception is raised.